Scoped clients
